Move ResultScroller controller plugin to standalone service.
